Rename containerId to imageId
Previously, we called the command line parameter that referred to the docker container image as "containerId". This is ambiguous, because it could also refer to a contaniner's id. Instead, it is more accurate to name this parameter as imageId. This commit makes that change, and also notes in the help documentation that it can refer to an image tag.