Keep publishing items even if previous item fails
Fix for Issue #64 (Failure when Publishing multiple items).

If an item can't be published, still try to publish subsequent items.

An error message will be printed for every unsuccessful item. In
addition, exit code will be FATAL (1) if any item had a fatal error
publishing, or will be RETRYABLE (2) if all failed items are retryable.